+++ docs/changelog.md
@@ -26,6 +26,7 @@
 - Added `gaussian unrestricted` and `maximum` optimization methods to `ScanningOptimizeLogic`
 - Supported pyvisa for coherent obis laser
 - Added Magnet Toolchain
+- Added derived period and pi-pulse duration to pulsed sine-fit results
 
 ### Other
 - Displaying Qudi logs in the test workflow upon test failure